# Bug 1815579 - Improve performance of image loading in tab items