--- a/fenix/app/src/main/java/org/mozilla/fenix/compose/TabThumbnail.kt
+++ b/fenix/app/src/main/java/org/mozilla/fenix/compose/TabThumbnail.kt
@@ -5,7 +5,6 @@
 package org.mozilla.fenix.compose
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -15,25 +14,23 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import mozilla.components.browser.icons.compose.Loader
-import mozilla.components.browser.icons.compose.Placeholder
-import mozilla.components.browser.icons.compose.WithIcon
-import org.mozilla.fenix.components.components
+import mozilla.components.browser.state.state.TabSessionState
+import mozilla.components.browser.state.state.createTab
 import org.mozilla.fenix.theme.FirefoxTheme
 
 private const val THUMBNAIL_SIZE = 108
 private const val FALLBACK_ICON_SIZE = 36
 
 /**
- * Card which will display a thumbnail. If a thumbnail is not available for [url], the favicon
+ * Thumbnail belonging to a [tab]. If a thumbnail is not available, the favicon
  * will be displayed until the thumbnail is loaded.
  *
- * @param url Url to display thumbnail for.
- * @param key Key used to remember the thumbnail for future compositions.
+ * @param tab The given [TabSessionState] to render a thumbnail for.
  * @param size [Dp] size of the thumbnail.
  * @param backgroundColor [Color] used for the background of the favicon.
  * @param modifier [Modifier] used to draw the image content.
@@ -43,12 +40,12 @@ private const val FALLBACK_ICON_SIZE = 36
  * @param alignment [Alignment] used to draw the image content.
  */
 @Composable
-fun ThumbnailCard(
-    url: String,
-    key: String,
+@Suppress("LongParameterList")
+fun TabThumbnail(
+    tab: TabSessionState,
+    modifier: Modifier = Modifier,
     size: Dp = THUMBNAIL_SIZE.dp,
     backgroundColor: Color = FirefoxTheme.colors.layer2,
-    modifier: Modifier = Modifier,
     contentDescription: String? = null,
     contentScale: ContentScale = ContentScale.FillWidth,
     alignment: Alignment = Alignment.TopCenter,
@@ -58,31 +55,32 @@ fun ThumbnailCard(
         backgroundColor = backgroundColor,
     ) {
         ThumbnailImage(
-            key = key,
+            key = tab.id,
             size = size,
             modifier = modifier,
             contentScale = contentScale,
             alignment = alignment,
         ) {
-            components.core.icons.Loader(url) {
-                Placeholder {
-                    Box(modifier = Modifier.background(color = FirefoxTheme.colors.layer3))
-                }
-
-                WithIcon { icon ->
-                    Box(
-                        modifier = Modifier.size(FALLBACK_ICON_SIZE.dp),
-                        contentAlignment = Alignment.Center,
-                    ) {
-                        Image(
-                            painter = icon.painter,
-                            contentDescription = contentDescription,
-                            modifier = Modifier
-                                .size(FALLBACK_ICON_SIZE.dp)
-                                .clip(RoundedCornerShape(8.dp)),
-                            contentScale = contentScale,
-                        )
-                    }
+            Box(
+                modifier = Modifier.size(FALLBACK_ICON_SIZE.dp),
+                contentAlignment = Alignment.Center,
+            ) {
+                val icon = tab.content.icon
+                if (icon != null) {
+                    icon.prepareToDraw()
+                    Image(
+                        bitmap = icon.asImageBitmap(),
+                        contentDescription = contentDescription,
+                        modifier = Modifier
+                            .size(FALLBACK_ICON_SIZE.dp)
+                            .clip(RoundedCornerShape(8.dp)),
+                        contentScale = contentScale,
+                    )
+                } else {
+                    Favicon(
+                        url = tab.content.url,
+                        size = FALLBACK_ICON_SIZE.dp,
+                    )
                 }
             }
         }
@@ -93,11 +91,10 @@ fun ThumbnailCard(
 @Composable
 private fun ThumbnailCardPreview() {
     FirefoxTheme {
-        ThumbnailCard(
-            url = "https://mozilla.com",
-            key = "123",
+        TabThumbnail(
+            tab = createTab(url = "www.mozilla.com", title = "Mozilla"),
             modifier = Modifier
-                .size(THUMBNAIL_SIZE.dp)
+                .size(THUMBNAIL_SIZE.dp, 80.dp)
                 .clip(RoundedCornerShape(8.dp)),
         )
     }