--- a/fenix/app/src/main/java/org/mozilla/fenix/home/recenttabs/view/RecentTabs.kt
+++ b/fenix/app/src/main/java/org/mozilla/fenix/home/recenttabs/view/RecentTabs.kt
@@ -56,7 +56,7 @@ import org.mozilla.fenix.components.components
 import org.mozilla.fenix.compose.ContextualMenu
 import org.mozilla.fenix.compose.Image
 import org.mozilla.fenix.compose.MenuItem
-import org.mozilla.fenix.compose.ThumbnailCard
+import org.mozilla.fenix.compose.TabThumbnail
 import org.mozilla.fenix.compose.annotation.LightDarkPreview
 import org.mozilla.fenix.compose.inComposePreview
 import org.mozilla.fenix.home.recenttabs.RecentTab
@@ -228,9 +228,8 @@ fun RecentTabImage(
                 contentScale = ContentScale.Crop,
             )
         }
-        else -> ThumbnailCard(
-            url = tab.state.content.url,
-            key = tab.state.id,
+        else -> TabThumbnail(
+            tab = tab.state,
             modifier = modifier,
             contentScale = contentScale,
         )