### Findings

#### Thumbnail bitmap retention
Thumbnail bitmaps were seemingly fetched, retained, and re-fetched. They were garbage collected when the Tabs Tray was dismissed, but violently scrolling up and down would explode the in-app memory.
  - My solution for this was to add a bitmap recycle call to `ThumbnailImage` in a `DisposableEffect` block.

#### The Favicon composable
Relying on the `Favicon` Composable when the icon was available via `TabsessionState.content.icon` was causing unnecessary icon fetches inside `TabGridItem` AND `ThumbnailCard`. 
  - I ended up needing to differentiate use cases for the original `ThumbnailCard`; one via Recent Synced Tabs (where `TabSessionState` isn't available) and one where `TabSessionState` is the driving data object. `TabThumbnail` is driven via `TabSessionState` so that the passed-in bitmap of `tab.content.icon` can be used instead of needlessly re-fetching the favicon.
  - The favicon (whether via `Favicon` or `TabsessionState.content.icon`) was also being loaded & rendered before the thumbnail was loaded AND they were existing simultaneously as the thumbnail (the thumbnail was being painted on top of the favicon). I refactored `ThumbnailImage` to have a `fallbackContent` Composable to render the favicon for a tab ONLY when the thumbnail could not be fetched. 
  - I also added an optional painter to `FaviconListitem` to cover this use case. This will aid in the performance of loading inactive tabs that have their icons still on disk.

#### Lazy list/grid composables
I noticed that some of the thumbnail fetching was built around the idea that the `LazyColumn` and `LazyGrid` Composables functioned 1:1 as a `RecyclerView`. I discovered this is not the case; Composables are NOT reused like Views are in the XML analog. We should not be coding our item Composables to take lazy loading into account but instead let their parent Composables handle this behind the scenes. I refactored the logic around fetching the thumbnails around a changing tab key string, since the key was never supposed to be changing in the first place. I also changed the bitmap `remember` to `rememberSaveable` as that will allow it to be preserved via app configuration changes. 

#### Bitmap fetching + side effect coroutines
One thing I'm not 100% certain on is using the coroutines from [side effects](https://developer.android.com/jetpack/compose/side-effects) to perform intense, time sensitive operations, like bitmap fetching in this situation. One hypothesis we had during investigation was the bitmap fetching coroutine in `LaunchedEffect` was launching _several seconds_ from when it should be. I opted to directly use `rememberCoroutineScope` instead to ensure the bitmap fetching coroutines are launched and de-scoped correctly.  

[One resource I used](https://developer.android.com/jetpack/compose/graphics/images/optimization)

### Profiles
- [Perf changes w/ 50 tabs, big fling down, big fling up](https://share.firefox.dev/3Jz8niy) (taken directly from the Firefox profiler)
- [Perf changes - open tabs tray w/ 50 tabs](https://share.firefox.dev/3NlD2RE) (taken via Android Studio profiler)

I tried to take more extensive profiles, but the tool in Android Studio was crashing in this use flow when capturing the trace.

### UI test comparison on Pixel 2
| XML time (run time, total) | Compose time before changes (run time, total) | Compose time after changes (run time, total) |
| -- | -- | -- |
| 3m 44s, 6m 51s | 3m 30s, 7m 13s | 3m 31s, 6m 39s |

### Videos (captured on Pixel 5, non-release build)
| XML (main) | Compose (main) | Compose (with changes) |
| -- | -- | -- |
| https://github.com/mozilla-mobile/firefox-android/assets/87384386/6febaae5-c6fd-4b4f-8989-11ae2e85a2a2 | https://github.com/mozilla-mobile/firefox-android/assets/87384386/64185033-4fc6-4a04-a5e1-9436c92e2a0e | https://github.com/mozilla-mobile/firefox-android/assets/87384386/87cf04ec-6923-47f3-997d-d9e82647dfa6 |





This is visual anecdotal evidence, but I still wanted to provide a "look and feel" comparison.

### Pull Request checklist
<!-- Before submitting the PR, please address each item -->
- [ ] **Quality**: This PR builds and passes detekt/ktlint checks (A pre-push hook is recommended)
- [ ] **Tests**: This PR includes thorough tests or an explanation of why it does not
- [ ] **Changelog**: This PR includes [a changelog entry](https://github.com/mozilla-mobile/firefox-android/blob/main/docs/changelog.md) or does not need one
- [ ] **Accessibility**: The code in this PR follows [accessibility best practices](https://github.com/mozilla-mobile/firefox-android/blob/main/docs/shared/android/accessibility_guide.md) or does not include any user facing features

### After merge
- **Breaking Changes**: If this is a breaking Android Components change, please push a draft PR on [Reference Browser](https://github.com/mozilla-mobile/reference-browser) to address the breaking issues.

### To download an APK when reviewing a PR (after all CI tasks finished running):
1. Click on `Checks` at the top of the PR page.
2. Click on the `firefoxci-taskcluster` group on the left to expand all tasks.
3. Click on the `build-apk-{fenix,focus,klar}-debug` task you're interested in.
4. Click on `View task in Taskcluster` in the new `DETAILS` section.
5. The APK links should be on the right side of the screen, named for each CPU architecture.
















### GitHub Automation
https://bugzilla.mozilla.org/show_bug.cgi?id=1815579