@jonalmeida 2023-06-08
> I performed the tests above by running the Espresso tests for the XML and Compose Tabs Trays. For the XML runs, I disabled the two tests that have not been converted to Compose yet.

Are we only comparing performance to the XML version? Shouldn't we also run these tests without your changes as well to see the difference with and without the diff?

Are there any links to the profiles that were gathered in the investigation? In the top-right corner, you can click the 'Upload local profile' to get a shareable link which can be added to the bugzilla or this PR.