@MozillaNoah 2023-06-14
> Are we only comparing performance to the XML version? Shouldn't we also run these tests without your changes as well to see the difference with and without the diff?

I think that makes sense. Matt and I are gonna pair to figure out some more wins here, but I'll have to rebase this branch and re-grab results anyway, so I'll add another column for `Compose before changes`.

> Are there any links to the profiles that were gathered in the investigation? In the top-right corner, you can click the 'Upload local profile' to get a shareable link which can be added to the bugzilla or this PR.

Top-right corner of which application? I'll upload some profiles after the next round of testing.