@MozillaNoah 2023-06-27
I may have found another performance win. [According to this article](https://developer.android.com/jetpack/compose/graphics/images/optimization#preparetodraw), we need to call `Bitmap`'s `prepareToDraw` function in order to signal to the GPU ahead of time that some images need to be rendered. A call like this would normally be handled by image loading libraries, but since we're working with our own homebrew situation here, we need to call it ourselves.

The most recent push contains these changes for `ThumbnailImage`, `TabGridItem`'s corner favicon, and the inactive tabs' use of `FaviconListitem`